backend/requirements.txt
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
crayons==0.4.0
cryptosteganography==0.8.4
exitstatus==2.5.0
Flask==3.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
numpy==2.0.0
opencv-python==4.10.0.84
piexif==1.1.3
pillow==10.3.0
pycryptodomex==3.20.0
python-dotenv==1.0.1
requests==2.32.3
setuptools==70.1.1
stegano==0.11.3
urllib3==2.2.3
Werkzeug==3.1.3
